feat(gateway): add access control functions
Note they will always revert right now as the actual sender of
the cross domain message will be the wrapper bridge contracts.
We need to either have the bridge contracts be inherited or
pass their address to the constructor of the other bridge in
addition to the address of the gateway.