refactor: attempt at simplifying the phases logic
Rely more on inputs from view functions to pass the phases.
Removal of the active dispute kits state in KlerosCore.
Allow changing the dispute kit phase from Drawing to Resolving inside DK.passPhase().
For safety, DK.passPhase() can reset its phase if it ever is not Resolving when Core is Staking.
Disabled the contract size limit on the Hardhat network because it's not a limitation on Arbitrum.